+++ pipcompilemulti/features/limit_envs.py
@@ -1,6 +1,12 @@
 """
-Limit ``.in`` files
-===================
+Limit environments
+==================
+
+.. warning::
+
+    This flag is deprecated and will be removed in future releases.
+    Use :ref:`limit-in-files` instead.
+
 
 By default ``pip-compile-multi`` compiles all ``.in`` files in ``requirements`` directory.
 To limit compilation to only a subset, use
@@ -19,16 +25,13 @@ For example, to compile one file under Python2.7 and another under Python3.6, ru
     Locking requirements/deps27.in to requirements/deps27.txt. References: []
     $ virtual-env36/bin/pip-compile-multi -n deps36
     Locking requirements/deps36.in to requirements/deps36.txt. References: []
-
-
-.. automodule:: pipcompilemulti.features.forbid_post
 """
 
-from pipcompilemulti.utils import recursive_refs
-from .base import BaseFeature, ClickOption
+from .base import ClickOption
+from .limit_in_paths import LimitInPaths
 
 
-class LimitEnvs(BaseFeature):
+class LimitEnvs(LimitInPaths):
     """Limit discovered environments to specified subset."""
 
     OPTION_NAME = 'include_names'
@@ -40,27 +43,15 @@ class LimitEnvs(BaseFeature):
                   'references. Can be supplied multiple times.',
     )
 
-    def __init__(self):
-        self._all_envs = None
+    def __init__(self, controller):
+        # pylint: disable=super-with-arguments
+        self._controller = controller
+        super(LimitEnvs, self).__init__()
 
     @property
     def direct_envs(self):
         """Set of environments included by command line options."""
-        return set(self.value or [])
-
-    def on_discover(self, env_confs):
-        """Save set of all (recursive) included environments."""
-        included_and_refs = self.direct_envs
-        if not self.direct_envs:
-            # No limit means all envs included:
-            self._all_envs = [env['name'] for env in env_confs]
-            return
-        for name in set(included_and_refs):
-            included_and_refs.update(
-                recursive_refs(env_confs, name)
-            )
-        self._all_envs = included_and_refs
-
-    def included(self, env_name):
-        """Whether environment is included directly or by reference."""
-        return env_name in self._all_envs
+        return set(
+            self._controller.compose_input_file_path(env_name)
+            for env_name in self.value or []
+        )